fix rk4 and change imu data type to doubles

+++ common/data_format.hpp
@@ -35,11 +35,11 @@ namespace ILLIXR {
 
 	struct imu_type : switchboard::event {
 		time_point time;
-		Eigen::Vector3f angular_v;
-		Eigen::Vector3f linear_a;
+		Eigen::Vector3d angular_v;
+		Eigen::Vector3d linear_a;
 		imu_type(time_point time_,
-				 Eigen::Vector3f angular_v_,
-				 Eigen::Vector3f linear_a_)
+				 Eigen::Vector3d angular_v_,
+				 Eigen::Vector3d linear_a_)
 			: time{time_}
 			, angular_v{angular_v_}
 			, linear_a{linear_a_}
